change optimizer, remove scheduler for now, use resnet18, unfreeze encoder weights

+++ src/training/train.py
@@ -4,7 +4,7 @@ from src.data.generate_token import load_tokens
 from src.data.data_loader import RSICDataset
 from src.utilities.utils import AverageMeter, accuracy, calculate_caption_lengths
 from src.data.encode_sequence import decode_sequence
-from src.utilities.plotting import plot_losses
+from src.utilities.plotting import plot_losses, write
 import torch
 from torch.utils.data import DataLoader, random_split
 from torchvision import transforms
@@ -16,10 +16,13 @@ from tqdm import tqdm
 import os
 import datetime
 import shutil
+import math
+from torch.utils.tensorboard import SummaryWriter
 
-def delete_folder_contents(folder_path):
+
+def remove_existing_saved_model(folder_path):
   """
-  Deletes all files and subfolders within a folder (use with caution).
+  Deletes existing best model saved within a folder (use with caution).
 
   Args:
       folder_path (str): The path to the folder to be emptied.
@@ -27,10 +30,9 @@ def delete_folder_contents(folder_path):
 
   for filename in os.listdir(folder_path):
     file_path = os.path.join(folder_path, filename)
-    if os.path.isfile(file_path):
+    if os.path.isfile(file_path) and "best_model" in filename:
       os.remove(file_path)
-    elif os.path.isdir(file_path):
-      shutil.rmtree(file_path, ignore_errors=True)  # Handle potential errors
+
 
 
 def train_loop(dataloader_train, optimizer, cnn, lstm, train_losses, cross_entropy_loss, word_indexer):
@@ -57,7 +59,7 @@ def train_loop(dataloader_train, optimizer, cnn, lstm, train_losses, cross_entro
     train_losses.append(losses.avg)
 
 
-def validation_loop(dataloader_val, cnn, lstm, tokens_params, val_losses, cross_entropy_loss, word_indexer):
+def validation_loop(dataloader_val, cnn, lstm, tokens_params, val_losses, cross_entropy_loss, word_indexer, show_example=False):
     losses_v = AverageMeter()
     with torch.no_grad():
         for image, descriptions, _, _ in tqdm(dataloader_val):
@@ -82,8 +84,9 @@ def validation_loop(dataloader_val, cnn, lstm, tokens_params, val_losses, cross_
             word_idxs = torch.max(preds, dim=1)[1].tolist()
             initial_text = decode_sequence(descriptions.tolist(), tokens_params['tokenizer'])
             reconstructed_text = decode_sequence([word_idxs], tokens_params['tokenizer'])
-        print("\n", initial_text)
-        print(reconstructed_text, "\n") 
+        if show_example:
+            print("\n", initial_text)
+            print(reconstructed_text, "\n") 
         val_losses.append(losses_v.avg)
 
 
@@ -99,7 +102,7 @@ def launch_training(image_dir_path, train_captions_path, all_captions_path, toke
 
     arch = config['architecture']
 
-    if arch in ["alexnet", "vgg", "resnet"]:
+    if arch in ["alexnet", "vgg", "resnet18"]:
         input_size = (224, 224)
     elif arch == "googlelenet":
         input_size = (299, 299)
@@ -126,8 +129,8 @@ def launch_training(image_dir_path, train_captions_path, all_captions_path, toke
     # define models
     cnn = CNN(architecture=arch, freeze_weights=config['freeze_weights'], pretrained=config['pretrained'])
     lstm = LSTM(vocabulary_size=tokens_params['vocab_size'], encoder_dim=cnn.dim)
-    optimizer = optim.Adam(lstm.parameters(), lr=config['lr'])
-    scheduler = optim.lr_scheduler.StepLR(optimizer, 1)
+    parameters = list(cnn.parameters()) + list(lstm.parameters())
+    optimizer = optim.SGD(parameters, lr=config['lr'])
     cross_entropy_loss = nn.CrossEntropyLoss().cuda()
     epochs = config['epochs']
 
@@ -136,8 +139,6 @@ def launch_training(image_dir_path, train_captions_path, all_captions_path, toke
 
     # Get current date and time
     now = datetime.datetime.now()
-
-    # Access different components
     date_str = now.strftime("%Y-%m-%d")  # Format: YYYY-MM-DD
     time_str = now.strftime("%H:%M:%S")  # Format: HH:MM:SS
 
@@ -145,29 +146,39 @@ def launch_training(image_dir_path, train_captions_path, all_captions_path, toke
     best_model_folder = os.path.join(best_models_path, best_model_folder_name)
     os.makedirs(best_model_folder, exist_ok=True)
 
-    for epoch in range(1, epochs + 1):
-        print(f"Epoch: {epoch}/{epochs}")
-        scheduler.step()
+    best_loss = math.inf
+
+    show_plot = config['show_loss_plot']
+    
+    path_tensor_logs = os.path.join(best_model_folder, "tensor_logs")
+    writer = SummaryWriter(path_tensor_logs)
+
+    for epoch in range(1, epochs + 1): # start the training
+        print(f"\n\n=========> Epoch: {epoch}/{epochs} <=========")
         train_loop(train_loader, optimizer, cnn, lstm, train_losses, cross_entropy_loss, word_indexer)
-        validation_loop(val_loader, cnn, lstm, tokens_params, val_losses, cross_entropy_loss, word_indexer)
+        validation_loop(val_loader, cnn, lstm, tokens_params, val_losses, cross_entropy_loss, word_indexer, show_example=config['show_example'])
+        writer.add_scalars("Losses", {"train": train_losses[epoch - 1], "validation": val_losses[epoch - 1]}, epoch - 1)
+
         if epoch > 1:
-            plot_losses(train_losses, val_losses)
-            if val_losses[-1] < val_losses[-2]:
-                model_file = f'best_model_epoch-{str(epoch)}_loss-{val_losses[-1]}_.pth'
+            plot_losses(train_losses, val_losses, best_model_folder, show_plot=show_plot)
+            if val_losses[-1] < best_loss:
+                best_loss = val_losses[-1]
+                model_file = f'best_model_epoch-{str(epoch)}_loss-{best_loss}_.pth'
                 # check if there is a model already saved 
-                delete_folder_contents(best_model_folder)
+                remove_existing_saved_model(best_model_folder)
                 torch.save(lstm.state_dict(), os.path.join(best_model_folder, model_file))
-                # save json folder also
+                # save json configuration also
                 with open(os.path.join(best_model_folder, "config.json"), "w") as f:
                     json.dump(config, f, indent=4)
                 print('Saved model to ' + model_file)
-
+    writer.close()
 if __name__ == "__main__":
     # load data here ..
-    image_dir_path = r"/Users/sergiosuzerainosson/Documents/project/universite_project/s4/modelisation_vision/image-caption-generation/data/interim/images"
-    all_captions_path = r"/Users/sergiosuzerainosson/Documents/project/universite_project/s4/modelisation_vision/image-caption-generation/data/interim/captions_all.json"
-    train_captions_path = r"/Users/sergiosuzerainosson/Documents/project/universite_project/s4/modelisation_vision/image-caption-generation/data/interim/captions_200_train.json"
-    tokens_path = r"/Users/sergiosuzerainosson/Documents/project/universite_project/s4/modelisation_vision/image-caption-generation/data/processed/tokenizer.p"
-    path_config = r"/Users/sergiosuzerainosson/Documents/project/universite_project/s4/modelisation_vision/image-caption-generation/src/configuration/config.json"
-    best_models_path = r"/Users/sergiosuzerainosson/Documents/project/universite_project/s4/modelisation_vision/image-caption-generation/models_"
+    root = r"/Users/sergiosuzerainosson/Documents/project/universite_project/s4/modelisation_vision/image-caption-generation/"
+    image_dir_path = root + r"data/interim/images"
+    all_captions_path = root + r"data/interim/captions_all.json"
+    train_captions_path = root + r"data/interim/captions_200_train.json"
+    tokens_path = root + r"data/processed/tokenizer.p"
+    path_config = root + r"src/configuration/config.json"
+    best_models_path = root + r"models_"
     launch_training(image_dir_path, train_captions_path, all_captions_path, tokens_path, path_config, best_models_path)
